+++ src/Bridge/Postgres/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+0.9
+---
+ * Introduce `StoreFactory`
+ * Introduce a `lang` argument to configure the lang used in embeddings
+
 0.1
 ---